Starting to migrate common code to the base class
