+++ simkratos/CFD/kratos_CFD_wrapper.py
@@ -160,117 +160,6 @@ class CFDWrapper(KratosWrapper):
         self.setSolutionStepVariable1D(data, node, "FLAG_VARIABLE")
         self.setSolutionStepVariable1D(data, node, "IS_STRUCTURE")
 
-    def exportKratosNodes(self, src, dst, group):
-        """ Parses all kratos nodes to simphony points
-
-        Iterates over all nodes in the kratos mesh (src) and
-        converts them to simphony points (dst). While doing this operation
-        any node/point that has not currently been mapped will have his uuid
-        added in the 'id_map' of the wrapper
-
-        """
-
-        for node in src.GetNodes(group):
-
-            data = {}
-
-            self.getNodalData(data, node)
-
-            point_uid = None
-
-            if node.Id not in self.id_to_uuid_node_map:
-
-                point = SPoint(
-                    coordinates=(node.X, node.Y, node.Z),
-                    data=DataContainer(data),
-                    uid=point_uid
-                )
-
-                pid = dst.add([point])
-
-                self.id_to_uuid_node_map[node.Id] = pid[0]
-
-            else:
-
-                point = dst.get(uid=self.id_to_uuid_node_map[node.Id])
-
-                # iterate over the correct data
-                point.data = DataContainer(data)
-
-                dst.update([point])
-
-    def exportKratosElements(self, src, dst, group):
-        """ Parses all kratos elements to simphony cells
-
-        Iterates over all elements in the kratos mesh (src) and
-        converts them to simphony cells (dst). While doing this operation
-        any element/cell that has not currently been mapped will have his uuid
-        added in the 'id_map' of the wrapper
-
-        """
-
-        for element in src.GetElements(group):
-
-            element_uid = None
-
-            if element.Id not in self.id_to_uuid_element_map:
-
-                point_list = [
-                    self.id_to_uuid_node_map[pointl.Id]
-                    for pointl in element.GetNodes()
-                ]
-
-                cell = SCell(
-                    points=point_list,
-                    uid=element_uid
-                )
-
-                cid = dst.add([cell])
-
-                self.id_to_uuid_element_map[element.Id] = cid[0]
-
-            else:
-
-                # No data is stored in the element yet
-
-                pass
-
-    def exportKratosConditions(self, src, dst, group):
-        """ Parses all kratos conditions to simphony faces
-
-        Iterates over all conditions in the kratos mesh (src) and
-        converts them to simphony faces (dst). While doing this operation
-        any condition/face that has not currently been mapped will have
-        his uuid added in the 'id_map' of the wrapper
-
-        """
-
-        for condition in src.GetConditions(group):
-
-            condition_uid = None
-
-            if condition.Id not in self.id_to_uuid_condition_map:
-
-                point_list = [
-                    self.id_to_uuid_node_map[point.Id]
-                    for point in condition.GetNodes()
-                ]
-
-                face = SFace(
-                    points=point_list,
-                    uid=condition_uid
-                )
-
-                fid = dst.add([face])
-
-                self.id_to_uuid_condition_map[condition.Id] = fid[0]
-
-            else:
-
-                # No data is stored in the condition yet
-
-                pass
-
     def exportKratosDof(self, src, dst, group):
         """ Sets the Dof information for the appropiate points
